requirements.txt
boto3
Pillow
pymongo
gridfs
pytest
moto
ultralytics
